services/backend/pyproject.toml
[project]
name = "images-server"
version = "0.1.0"
description = "Image upload server with Docker and Nginx"
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pillow (>=11.2.1,<12.0.0)",
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"